aurnova-agent-template/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aurnova-agent-core"
version = "0.1.0"
description = "Aurnova Agent Core — base agent template for MSAI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.24",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*", "llm*", "tools*", "memory*", "orchestrator*", "eval*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
